v1

+++ BlackScholes/FSImpliedVol.py
@@ -24,8 +24,8 @@ class ImpliedVolCalculatorFS:
 
     def FSImpliedVol(self):
         # Conversion des listes en tenseurs GPU
-        k_values_tensor = torch.tensor(self.k_values, device=CONFIG.device, dtype=torch.float64)
-        T2_tensor = torch.tensor(self.T2, device=CONFIG.device, dtype=torch.float64)
+        k_values_tensor = torch.tensor(self.k_values, device=CONFIG.device, dtype=torch.float32)
+        T2_tensor = torch.tensor(self.T2, device=CONFIG.device, dtype=torch.float32)
 
         # Initialisation des paramètres optimisables (theta utilisé pour sigma implicite)
         theta = torch.tensor([0.2] * len(self.k_values), device=CONFIG.device, requires_grad=True)
