+++ BlackScholes/ForwardStartBlackScholes.py
@@ -4,11 +4,11 @@ from config import CONFIG
 
 class FSBlackScholes:
     def __init__(self, S0, k, T1, T2, r, sigma):
-        self.S0 = torch.tensor([S0], device=CONFIG.device, dtype=torch.float64, requires_grad=True)
-        self.k = torch.tensor(k, device=CONFIG.device,dtype=torch.float64)
-        self.T1 = torch.tensor([T1],device=CONFIG.device, dtype=torch.float64)
-        self.T2 = torch.tensor(T2,device=CONFIG.device, dtype=torch.float64)
-        self.r = torch.tensor([r],device=CONFIG.device, dtype=torch.float64)
+        self.S0 = torch.tensor([S0], device=CONFIG.device, dtype=torch.float32, requires_grad=True)
+        self.k = torch.tensor(k, device=CONFIG.device,dtype=torch.float32)
+        self.T1 = torch.tensor([T1],device=CONFIG.device, dtype=torch.float32)
+        self.T2 = torch.tensor(T2,device=CONFIG.device, dtype=torch.float32)
+        self.r = torch.tensor([r],device=CONFIG.device, dtype=torch.float32)
         self.sigma = sigma
 
     def _d1_d2(self):
